Refactor: Restructure into couchdb3.sync and couchdb3.aio subpackages
