+++ src/couchdb3/sync/server.py
@@ -5,14 +5,14 @@ import httpx
 
 from .base import Base
 from .database import Database
-from .exceptions import (
+from ..exceptions import (
     ConflictError,
     CouchDBError,
     NotFoundError,
     ProxySchemeComplianceError,
     UserIDComplianceError,
 )
-from .utils import (
+from ..utils import (
     DEFAULT_TIMEOUT,
     rm_nones_from_dict,
     user_name_to_id,
