+++ tests/test_async_database.py
@@ -5,10 +5,9 @@ import random
 import string
 import unittest
 
-from couchdb3.async_database import AsyncDatabase, AsyncPartition
-from couchdb3.async_server import AsyncServer
+from couchdb3.aio import AsyncDatabase, AsyncPartition, AsyncServer
 from couchdb3.document import AttachmentDocument, Document
-from couchdb3.server import Server
+from couchdb3.sync import Server
 from couchdb3.utils import MimeTypeEnum, user_name_to_id
 from couchdb3.view import ViewResult, ViewRow
 from tests.credentials import (
