+++ tests/test_partitioned_database.py
@@ -3,8 +3,7 @@
 import atexit
 import unittest
 
-from couchdb3.database import Database
-from couchdb3.server import Server
+from couchdb3.sync import Database, Server
 from tests.credentials import (
     COUCHDB0_URL,
     COUCHDB_PASSWORD,
